chore: symbol in separate file
