+++ src/utils.ts
@@ -1,7 +1,8 @@
+export { RESET } from './utils/constants'
 export { useUpdateAtom } from './utils/useUpdateAtom'
 export { useAtomValue } from './utils/useAtomValue'
 export { atomWithReset } from './utils/atomWithReset'
-export { useResetAtom, RESET } from './utils/useResetAtom'
+export { useResetAtom } from './utils/useResetAtom'
 export { useReducerAtom } from './utils/useReducerAtom'
 export { atomWithReducer } from './utils/atomWithReducer'
 export { atomFamily } from './utils/atomFamily'
