+++ src/utils/useResetAtom.ts
@@ -2,7 +2,7 @@ import { useCallback, useContext } from 'react'
 import { SECRET_INTERNAL_getStoreContext as getStoreContext } from 'jotai'
 import type { WritableAtom } from 'jotai'
 
-export const RESET = Symbol()
+import { RESET } from './constants'
 
 export function useResetAtom<Value>(anAtom: WritableAtom<Value, typeof RESET>) {
   const StoreContext = getStoreContext(anAtom.scope)
